refactor: move replicate API helpers into separate module
